changes primarily to styling of productpages in the shop. a step further towards the final outcome

+++ wp-content/plugins/wp-e-commerce/wpsc-merchants/GoogleCheckout-XML.php
@@ -1,13 +1,13 @@
 <?php
 
-require_once('library/googlecart.php');
-require_once('library/googleitem.php');
-require_once('library/googleshipping.php');
-require_once('library/googletax.php');
-require_once('library/googleresponse.php');
-require_once('library/googlemerchantcalculations.php');
-require_once('library/googleresult.php');
-require_once('library/googlerequest.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googlecart.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googleitem.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googleshipping.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googletax.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googleresponse.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googlemerchantcalculations.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googleresult.php');
+require_once(WPSC_FILE_PATH.'/wpsc-merchants/library/googlerequest.php');
 
 
 $nzshpcrt_gateways[$num]['name'] = 'Google Checkout';
@@ -47,17 +47,19 @@ function gateway_google($fromcheckout = false){
 		$update = $wpdb->query($sql);
 		$sql = "SELECT `id` FROM `".WPSC_TABLE_PURCHASE_LOGS."` WHERE sessionid=".$_SESSION['wpsc_sessionid'];
 		$purchase_log_id = $wpdb->get_var($sql);
-		$sql = "DELETE FROM  `".WPSC_TABLE_CART_CONTENTS."` WHERE purchaseid = ".$purchase_log_id;
-		$wpdb->query($sql);
-		$wpsc_cart->save_to_db($purchase_log_id);
+		if( !empty($purchase_log_id) ){
+			$sql = "DELETE FROM  `".WPSC_TABLE_CART_CONTENTS."` WHERE purchaseid = ".$purchase_log_id;
+			$wpdb->query($sql);
+		}
 		if(! $update){
 			$sql = "INSERT INTO `".WPSC_TABLE_PURCHASE_LOGS."` (`totalprice`,`statusno`, `sessionid`, `user_ID`, `date`, `gateway`, `billing_country`,`shipping_country`, `base_shipping`,`shipping_method`, `shipping_option`, `plugin_version`, `discount_value`, `discount_data`) VALUES ('$total' ,'0', '".$_SESSION['wpsc_sessionid']."', '".(int)$user_ID."', UNIX_TIMESTAMP(), 'google', '{$wpsc_cart->delivery_country}', '{$wpsc_cart->selected_country}', '{$base_shipping}', '".$wpsc_cart->selected_shipping_method."', '".$wpsc_cart->selected_shipping_option."', '".WPSC_VERSION."', '{$wpsc_cart->coupons_amount}','{$wpsc_cart->coupons_name}')";
 			$wpdb->query($sql);
 			$sql = "SELECT `id` FROM `".WPSC_TABLE_PURCHASE_LOGS."` WHERE sessionid=".$_SESSION['wpsc_sessionid'];
 			$purchase_log_id = $wpdb->get_var($sql);
-			$wpsc_cart->save_to_db($purchase_log_id);
-		}	
 		
+		}	
+		$wpsc_cart->save_to_db($purchase_log_id);
+
 		if(get_option('permalink_structure') != '') {
 			$separator = "?";
 		} else {
@@ -83,7 +85,6 @@ function gateway_google($fromcheckout = false){
 	$returnURL =  $transact_url.$separator."sessionid=".$sessionid."&gateway=google";
 	$cart->SetContinueShoppingUrl($returnURL);
 	$cart->SetEditCartUrl(get_option('shopping_cart_url'));
-	$no=1;
 	//new item code
 	$no = 0;
 	//google prohibited items not implemented
@@ -129,7 +130,7 @@ function gateway_google($fromcheckout = false){
 	// Add shipping options
 	if(wpsc_uses_shipping() && $google_currency_shipping >0 ){
 		$Gfilter = new GoogleShippingFilters();
-		$google_checkout_shipping=get_option("google_shipping_country");
+		$google_checkout_shipping = get_option("google_shipping_country");
 		$googleshippingcountries = count($google_checkout_shipping);
 		if($googleshippingcountries == 242){
 			$Gfilter->SetAllowedWorldArea(true);
@@ -148,7 +149,7 @@ function gateway_google($fromcheckout = false){
 		}
 		
 		$Gfilter->SetAllowUsPoBox(false);
-		$ship_1 = new GoogleFlatRateShipping('Flat Rate Shipping', $google_currency_shipping);
+		$ship_1 = new GoogleFlatRateShipping($wpsc_cart->selected_shipping_method, $google_currency_shipping);
 		$ship_1->AddShippingRestrictions($Gfilter);
 		$cart->AddShipping($ship_1);
 	}
